Add logging functionality and base interface for Pelmen module. Set up notifications to be sent to Telegram.

+++ modules/user/controllers/UserController.php
@@ -2,14 +2,19 @@
 
 namespace app\modules\user\controllers;
 
+use Yii;
 use app\modules\user\models\User;
 use yii\web\Controller;
+use app\modules\user\models\LoginForm;
 
 class UserController extends Controller
 {
     public function actionIndex()
     {
-        // Получаем всех пользователей с ролями
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(['/login']);
+        }
+
         $dataProvider = new \yii\data\ActiveDataProvider([
             'query' => User::find()->with('role'),
         ]);
@@ -21,9 +26,13 @@ class UserController extends Controller
 
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(['/login']);
+        }
+
         $model = new User();
 
-        if ($model->load(\Yii::$app->request->post()) && $model->save()) {
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['index']);
         }
 
@@ -32,16 +41,24 @@ class UserController extends Controller
 
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(['/login']);
+        }
+
         $user = User::findOne($id);
-        return $this->render('views', ['user' => $user]);
+        return $this->render('view', ['user' => $user]);
     }
 
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(['/login']);
+        }
+
         $user = User::findOne($id);
 
-        if ($user->load(\Yii::$app->request->post()) && $user->save()) {
-            return $this->redirect(['views', 'id' => $user->id]);
+        if ($user->load(Yii::$app->request->post()) && $user->save()) {
+            return $this->redirect(['view', 'id' => $user->id]);
         }
 
         return $this->render('update', ['model' => $user]);
@@ -49,7 +66,33 @@ class UserController extends Controller
 
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(['/login']);
+        }
+
         User::findOne($id)->delete();
         return $this->redirect(['index']);
+    }
+
+    public function actionLogin()
+    {
+        if (!Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
+        $model = new LoginForm();
+        if ($model->load(Yii::$app->request->post()) && $model->login()) {
+            return $this->goBack();
+        }
+
+        return $this->render('login', [
+            'model' => $model,
+        ]);
+    }
+
+    public function actionLogout()
+    {
+        Yii::$app->user->logout();
+        return $this->goHome();
     }
 }
